# Conflicts:
#	cmake/CMakeLists.txt
#	doc/src/Commands_fix.txt
#	doc/src/Packages_details.txt
#	doc/src/Packages_user.txt
#	doc/src/fix_nh.txt
#	doc/src/fixes.txt
#	doc/src/lammps.book
#	src/Makefile

+++ src/USER-PTM/compute_ptm_atom.h
@@ -35,7 +35,7 @@ class ComputePTMAtom : public Compute {
 
  private:
   int nmax;
-  int32_t input_flags;
+  smallint input_flags;
   double rmsd_threshold;
   class NeighList *list;
   double **output;
